+++ packages/moonstone/CHANGELOG.md
@@ -6,13 +6,14 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Deprecated
 
+- `moonstone/Button` prop `noAnimation`
+
 ### Added
 
 ### Changed
 
 - `moonstone/Marquee` to do less-costly calculations during measurement and optimized the applied styles
 - `moonstone/ExpandableList` to require a unique key for each object type data
-- `moonstone/Button` to deprecate prop `noAnimation`
 
 ### Fixed